Added adminController
